fixed create problem

+++ PROYECTO1_ED1/Models/ModeloPaciente.cs
@@ -20,7 +20,7 @@ namespace PROYECTO1_ED1.Models
         [Required]
         public int Edad { get; set; }
         [Required]
-        public int Teléfono { get; set; }
+        public long Teléfono { get; set; }
         [Required]
         public DateTime ÚltimaConsulta { get; set; }
         public DateTime PróximaConsulta { get; set; }
@@ -28,9 +28,7 @@ namespace PROYECTO1_ED1.Models
 
         public static bool Guardar(ModeloPaciente datos)
         {
-            
-
-            
+         
             Data.Instance.ÁrbolPacientes.Insert(datos);
             return true; 
         }
